+++ Raytracer/BaseClasses/Hitable.cs
@@ -1,14 +1,11 @@
 ﻿using Raytracer.Geometry;
 using Raytracer.Materials;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Raytracer.BaseClasses
 {
     public struct HitRecord
     {
-        public float T;
+        public double T;
         public Vec3 P;
         public Vec3 Normal;
         public Material Material;
@@ -16,6 +13,6 @@ namespace Raytracer.BaseClasses
 
     public abstract class Hitable
     {
-        public abstract bool Hit(Ray r, float tMin, float tMax, out HitRecord rec);
+        public abstract bool Hit(Ray r, double tMin, double tMax, out HitRecord rec);
     }
 }
